Unify "backend" and "embed" into one file
