+++ packages/react-devtools-core/package.json
@@ -2,7 +2,7 @@
   "name": "react-devtools-core",
   "version": "1.0.6",
   "description": "Use react-devtools outside of the browser",
-  "main": "./build/embed.js",
+  "main": "./build/backend.js",
   "repository": {
     "url": "https://github.com/facebook/react-devtools.git",
     "type": "git"
@@ -14,10 +14,10 @@
   ],
   "scripts": {
     "backend": "NODE_ENV=production ../../node_modules/.bin/webpack --config webpack.backend.js",
-    "embed": "NODE_ENV=production ../../node_modules/.bin/webpack --config webpack.embed.js",
+    "backend:watch": "NODE_ENV=production ../../node_modules/.bin/webpack --config webpack.backend.js --watch",
     "standalone": "NODE_ENV=production ../../node_modules/.bin/webpack --config webpack.standalone.js",
     "standalone:watch": "NODE_ENV=production ../../node_modules/.bin/webpack --config webpack.standalone.js --watch",
-    "build": "npm run backend && npm run embed && npm run standalone"
+    "build": "npm run backend && npm run standalone"
   },
   "author": "Jared Forsyth",
   "license": "BSD-3-Clause",
